requirements.txt
Django==3.1.2
gunicorn==20.0.4
django-ckeditor==6.0.0
mysqlclient==2.0.3
